pyproject.toml
[tool.poetry]
name = "okta-login"
version = "0.1.0"
description = ""
authors = ["Daniel Bajorek <danielbajorek@mail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
importlib-metadata = "^6.8.0"
requests = "^2.31.0"
flask = "2.3.3"
flask-cors = "^4.0.0"
pyopenssl = "^23.2.0"
flask-login = "^0.6.2"
okta-jwt-verifier = "^0.2.4"
werkzeug = "2.3.7"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
